# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API",
    "version": "1.0.0",
    "description": "API for managing contacts"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "summary": "Get all contacts",
        "tags": [
          "Contacts"
        ],
        "responses": {
          "200": {
            "description": "List of contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "tags": [
          "Contacts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "Conflict - Email already exists"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "summary": "Get contact by ID",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Contact ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing contact",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Contact ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Contact not found"
          },
          "409": {
            "description": "Conflict - Email already exists"
          }
        }
      },
      "delete": {
        "summary": "Delete a contact",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Contact ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted"
          },
          "404": {
            "description": "Contact not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Contact": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "favoriteColor",
          "birthday",
          "phone",
          "address"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the contact."
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the contact."
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the contact."
          },
          "email": {
            "type": "string",
            "description": "The email address of the contact."
          },
          "favoriteColor": {
            "type": "string",
            "description": "The favorite color of the contact."
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "description": "The birthday of the contact."
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the contact."
          },
          "address": {
            "type": "string",
            "description": "Address of the contact."
          }
        },
        "example": {
          "id": "60f7f9b2e1d3c0b1c8a9e123",
          "firstName": "Abigail",
          "lastName": "Johnson",
          "email": "abi.jo@example.com",
          "favoriteColor": "Blue",
          "birthday": "1990-01-01",
          "phone": "123-456-7890",
          "address": "123 Main St, Springfield"
        }
      },
      "ContactInput": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "favoriteColor",
          "birthday",
          "phone",
          "address"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "description": "The first name of the contact."
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the contact."
          },
          "email": {
            "type": "string",
            "description": "The email address of the contact."
          },
          "favoriteColor": {
            "type": "string",
            "description": "The favorite color of the contact."
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "description": "The birthday of the contact."
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the contact."
          },
          "address": {
            "type": "string",
            "description": "Address of the contact."
          }
        },
        "example": {
          "firstName": "Abigail",
          "lastName": "Johnson",
          "email": "abi.jo@example.com",
          "favoriteColor": "Blue",
          "birthday": "1990-01-01",
          "phone": "123-456-7890",
          "address": "123 Main St, Springfield"
        }
      },
      "ContactResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ContactInput"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "The auto-generated id of the contact."
              }
            }
          }
        ],
        "example": {
          "id": "60f7f9b2e1d3c0b1c8a9e123",
          "firstName": "Abigail",
          "lastName": "Johnson",
          "email": "abi.jo@example.com",
          "favoriteColor": "Blue",
          "birthday": "1990-01-01",
          "phone": "123-456-7890",
          "address": "123 Main St, Springfield"
        }
      }
    }
  },
  "tags": [
    {
      "name": "Contacts",
      "description": "API for managing contacts"
    }
  ]
}